Admin flag fail
So I was actually wrong. There is no admin flag like I thought there was . I added the bot to one of my groups and tested the admin changes and they didn't work. Did some digging and realized I was reading some other api docs. The way that you had previously mentioned which is checking to see if kicking the user works is the best way to go about avoiding sending the kick message and deleting the offending message. This is very hard for me to test locally , so sorry I had to bother you so much. It would be great if you could test this to make sure it is still working before deploying, but if you want to deploy to test thats fine as well I have made a groupchat to test it with